src/test/java/eu/trentorise/opendata/traceprov/test/ProvRefsTest.java
/*
 * Copyright 2015 Trento Rise.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package eu.trentorise.opendata.traceprov.test;

import eu.trentorise.opendata.traceprov.data.DcatMetadata;
import static eu.trentorise.opendata.traceprov.schema.ProvRefs.propertyRef;
import static org.junit.Assert.assertEquals;
import org.junit.Test;

/**
 *
 * @author David Leoni
 */
public class ProvRefsTest {
    
    @Test
    public void test(){
        assertEquals("catalog", propertyRef(DcatMetadata.class, "catalog"));
    }
}
